@@ -1,6 +1,6 @@
 {
   "name": "@igo2/common",
-  "version": "1.14.2",
+  "version": "1.15.0",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": ["igo"],
@@ -14,15 +14,19 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/material": "^14.1.2",
-    "@igo2/core": "^1.14.2",
-    "@igo2/utils": "^1.14.2",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@igo2/core": "^1.15.0",
+    "@igo2/utils": "^1.15.0",
     "scroll-into-view-if-needed": "^2.2.20",
     "typy": "^3.3.0"
   },
   "dependencies": {
     "tslib": "^2.4.0"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
