@@ -1,6 +1,6 @@
 {
   "name": "@igo2/core",
-  "version": "1.14.2",
+  "version": "1.15.0",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": ["igo"],
@@ -18,12 +18,17 @@
     "tslib": "^2.4.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^14.1.2",
-    "@angular/common": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@igo2/utils": "^1.14.2",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@igo2/utils": "^1.15.0",
     "rxjs": "^7.5.6",
     "ngx-indexed-db": "^11.0.2",
     "ngx-toastr": "^15.0.0"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
