@@ -1,6 +1,6 @@
 {
   "name": "@igo2/geo",
-  "version": "1.14.2",
+  "version": "1.15.0",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": ["igo"],
@@ -24,20 +24,22 @@
     "windows-1252": "^3.0.4"
   },
   "peerDependencies": {
-    "@angular/animations": "^14.1.2",
-    "@angular/common": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.2",
-    "@igo2/common": "^1.14.2",
-    "@igo2/core": "^1.14.2",
-    "@igo2/utils": "^1.14.2",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@igo2/common": "^1.15.0",
+    "@igo2/core": "^1.15.0",
+    "@igo2/utils": "^1.15.0",
     "flexsearch": "0.7.21",
     "ol": "^7.3.0",
     "proj4": "^2.8.0",
     "rxjs": "^7.5.6",
     "striptags": "^3.2.0",
-    "ngx-color-picker": "^13.0.0"
+    "ngx-color-picker": "^13.0.0",
+    "ngx-indexed-db": "^11.0.2"
   },
   "optionalDependencies": {
     "file-saver": "^2.0.2",
@@ -46,5 +48,8 @@
     "jszip": "^3.10.1",
     "moment": "^2.29.4",
     "stream": "^0.0.2"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
