@@ -86,4 +86,14 @@ export class AnalyticsService {
       this.paq.push(['trackEvent', category, action, name]);
     }
   }
-}
+
+  /**
+   * Function that tracks layers added to the map
+   */
+  public trackLayer(category: string, action: string, parameters: any){
+    if (this.options.provider === 'matomo')
+          this.paq.push(['trackEvent', category, action, parameters]);
+
+      }
+    }
+
