@@ -43,14 +43,19 @@ export class MessageService {
 
           if (activeMessageTranslation.textInterpolateParams) {
             Object.keys(activeMessageTranslation.textInterpolateParams).map(k => {
-              translatedTextInterpolateParams[k] =
-                this.languageService.translate.instant(activeMessageTranslation.textInterpolateParams[k]);
+              if (k) {
+                translatedTextInterpolateParams[k] =
+                  this.languageService.translate.instant(activeMessageTranslation.textInterpolateParams[k]);
+              }
             });
           }
           if (activeMessageTranslation.titleInterpolateParams) {
             Object.keys(activeMessageTranslation.titleInterpolateParams).map(k => {
-              translatedTitleInterpolateParams[k] =
-                this.languageService.translate.instant(activeMessageTranslation.titleInterpolateParams[k]);
+              if (k) {
+                translatedTitleInterpolateParams[k] =
+                  this.languageService.translate.instant(activeMessageTranslation.titleInterpolateParams[k]);
+              }
+
             });
           }
 
@@ -215,14 +220,18 @@ export class MessageService {
 
     if (textInterpolateParams) {
       Object.keys(textInterpolateParams).map(k => {
-        translatedTextInterpolateParams[k] =
-          this.languageService.translate.instant(textInterpolateParams[k]);
+        if (textInterpolateParams[k]) {
+          translatedTextInterpolateParams[k] =
+            this.languageService.translate.instant(textInterpolateParams[k]);
+        }
       });
     }
     if (titleInterpolateParams) {
       Object.keys(titleInterpolateParams).map(k => {
-        translatedTitlenterpolateParams[k] =
-          this.languageService.translate.instant(titleInterpolateParams[k]);
+        if (k) {
+          translatedTitlenterpolateParams[k] =
+            this.languageService.translate.instant(titleInterpolateParams[k]);
+        }
       });
     }
 
