@@ -14,10 +14,11 @@ import { EntityStore } from '@igo2/common';
 import type { EntityStateManager } from '@igo2/common';
 
 import {
+  addedChangeEmitter,
+  addedChangeGroupEmitter,
   Catalog,
   CatalogItem,
   CatalogItemGroup,
-  CatalogItemLayer,
   CatalogItemState,
   CatalogItemType
 } from '../shared';
@@ -90,18 +91,12 @@ export class CatalogBrowserGroupComponent implements OnInit, OnDestroy {
   /**
    * Event emitted when the add/remove button of the group is clicked
    */
-  @Output() addedChange = new EventEmitter<{
-    added: boolean;
-    group: CatalogItemGroup;
-  }>();
+  @Output() addedChange = new EventEmitter<addedChangeGroupEmitter>();
 
   /**
    * Event emitted when the add/remove button of a layer is clicked
    */
-  @Output() layerAddedChange = new EventEmitter<{
-    added: boolean;
-    layer: CatalogItemLayer;
-  }>();
+  @Output() layerAddedChange = new EventEmitter<addedChangeEmitter>();
 
   /**
    * @internal
@@ -147,8 +142,8 @@ export class CatalogBrowserGroupComponent implements OnInit, OnDestroy {
    * On toggle button click, emit the added change event
    * @internal
    */
-  onToggleClick() {
-    this.added$.value ? this.remove() : this.add();
+  onToggleClick(event: Event) {
+    this.added$.value ? this.remove(event) : this.add(event);
   }
 
   /**
@@ -166,7 +161,7 @@ export class CatalogBrowserGroupComponent implements OnInit, OnDestroy {
    * @internal
    * @param event Layer added change event
    */
-  onLayerAddedChange(event: { added: boolean; layer: CatalogItemLayer }) {
+  onLayerAddedChange(event: addedChangeEmitter) {
     this.layerAddedChange.emit(event);
     this.tryToggleGroup(event);
   }
@@ -174,22 +169,24 @@ export class CatalogBrowserGroupComponent implements OnInit, OnDestroy {
   /**
    * Emit added change event with added = true
    */
-  private add() {
+  private add(event: Event) {
     this.added$.next(true);
     this.addedChange.emit({
       added: true,
-      group: this.group
+      group: this.group,
+      event
     });
   }
 
   /**
    * Emit added change event with added = true
    */
-  private remove() {
+  private remove(event: Event) {
     this.added$.next(false);
     this.addedChange.emit({
       added: false,
-      group: this.group
+      group: this.group,
+      event
     });
   }
 
@@ -201,7 +198,7 @@ export class CatalogBrowserGroupComponent implements OnInit, OnDestroy {
    * If all the layers of the group added or removed, add or remove the group itself.
    * @param event The last layer added change event to occur
    */
-  private tryToggleGroup(event: { added: boolean; layer: CatalogItemLayer }) {
+  private tryToggleGroup(event: addedChangeEmitter) {
     const added = event.added;
     const layer = event.layer;
 
@@ -211,7 +208,7 @@ export class CatalogBrowserGroupComponent implements OnInit, OnDestroy {
       .map((item: CatalogItem) => this.state.get(item).added || false);
 
     if (layersAdded.every(value => value === added)) {
-      added ? this.add() : this.remove();
+      added ? this.add(event.event) : this.remove(event.event);
     } else if (this.added$.value === true) {
       this.added$.next(false);
     }
