@@ -29,9 +29,14 @@ import {
 import { EntityStore } from '@igo2/common';
 import { getTooltipsOfOlGeometry } from '../../measure';
 import OlOverlay from 'ol/Overlay';
+import Stroke from 'ol/style/Stroke';
+import Fill from 'ol/style/Fill';
+import Style from 'ol/style/Style';
+import Circle from 'ol/style/Circle';
 import { VectorSourceEvent as OlVectorSourceEvent } from 'ol/source/Vector';
 import { default as OlGeometry } from 'ol/geom/Geometry';
 import { QueryableDataSourceOptions } from '../../query';
+import { Media, MediaService } from '@igo2/core';
 
 
 @Component({
@@ -78,6 +83,11 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
    */
   @Input() map: IgoMap;
 
+  /**
+   * show hide save search result in layer button
+   */
+  @Input() saveSearchResultInLayer: boolean = false;
+
   @Input()
   get color() {
     return this._color;
@@ -94,16 +104,26 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
       String(layer.id).includes('igo-search-layer')
     );
   }
-
+  private mediaService$$: Subscription;
+  public isMobile: boolean = false;
   constructor(
     private layerService: LayerService,
     private dialog: MatDialog,
-    private dataSourceService: DataSourceService) {}
+    private dataSourceService: DataSourceService,
+    private mediaService: MediaService) {}
 
   /**
    * @internal
    */
   ngOnInit(): void {
+    // check the view if is mobile or not
+    this.mediaService$$ = this.mediaService.media$.subscribe(
+      (media: Media) => {
+        if(media === Media.Mobile) {
+          this.isMobile = true;
+        }
+      }
+    );
     if (this.layer.meta.dataType === 'Layer') {
       this.added =
         this.map.layers.findIndex(
@@ -122,6 +142,9 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
   ngOnDestroy() {
     this.resolution$$.unsubscribe();
     this.layers$$.unsubscribe();
+    if (this.mediaService$$) {
+      this.mediaService$$.unsubscribe();
+    }
   }
 
   /**
@@ -150,7 +173,7 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
           if (this.added) {
             this.remove();
           } else {
-            this.add();
+            this.add(event);
           }
         }
         this.isPreview$.next(false);
@@ -176,10 +199,10 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
     }
   }
 
-  private add() {
+  private add(event?: Event) {
     if (!this.added) {
       this.added = true;
-      this.addLayerToMap();
+      this.addLayerToMap(event);
     }
   }
 
@@ -195,7 +218,7 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
   /**
    * Emit added change event with added = true
    */
-  private addLayerToMap() {
+  private addLayerToMap(event?: Event) {
     if (this.map === undefined) {
       return;
     }
@@ -211,7 +234,12 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
     this.layersSubcriptions.push(
       this.layerService
         .createAsyncLayer(layerOptions)
-        .subscribe(layer => this.map.addLayer(layer))
+        .subscribe(layer => {
+          if (event.type === 'click') {
+            this.map.layersAddedByClick$.next([layer]);
+          }
+          this.map.addLayer(layer);
+        })
     );
   }
 
@@ -309,10 +337,33 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
   }
 
   createLayer(layerTitle: string, selectedFeature: SearchResult) {
-
     const activeStore: FeatureStore<Feature> = new FeatureStore<Feature>([], {
       map: this.map
     });
+
+    const styles = [
+      new Style({
+        image: new Circle({
+          radius: 5,
+          stroke: new Stroke({
+            width: 1,
+            color: 'rgba(143,7,7,1)'
+          }),
+          fill: new Fill({
+            color: 'rgba(143,7,7,1)'
+          })
+        })
+      }),
+      new Style({
+        stroke: new Stroke({
+          width: 1,
+          color: 'rgba(143,7,7,1)'
+        }),
+        fill: new Fill({
+          color: 'rgba(0, 0, 255, 0.1)',
+        }),
+      })
+    ];
 
     // set layer id
     let layerCounterID: number = 0;
@@ -345,6 +396,7 @@ export class SearchResultAddButtonComponent implements OnInit, OnDestroy{
                 }
               }
             },
+            style: styles,
             showInLayerList: true,
             exportable: true,
             workspace: {
