@@ -1,10 +1,24 @@
 import { Injectable } from '@angular/core';
 
-import { EntityRecord, EntityStore, EntityStoreFilterCustomFuncStrategy, EntityStoreStrategyFuncOptions } from '@igo2/common';
+import {
+  EntityRecord,
+  EntityStore,
+  EntityStoreFilterCustomFuncStrategy,
+  EntityStoreStrategyFuncOptions
+} from '@igo2/common';
 import { ConfigService, StorageService } from '@igo2/core';
-import { SearchResult, SearchSourceService, SearchSource, CommonVectorStyleOptions, FeatureWorkspace } from '@igo2/geo';
+import {
+  SearchResult,
+  SearchSourceService,
+  SearchSource,
+  CommonVectorStyleOptions,
+  FeatureWorkspace,
+  FeatureStore,
+  Feature
+} from '@igo2/geo';
 import { BehaviorSubject, Subscription } from 'rxjs';
 import { WorkspaceState } from '../workspace/workspace.state';
+import { MapState } from '../map';
 
 /**
  * Service that holds the state of the search module
@@ -13,6 +27,7 @@ import { WorkspaceState } from '../workspace/workspace.state';
   providedIn: 'root'
 })
 export class SearchState {
+  public searchLayerStores: FeatureStore<Feature>[] = [];
   public searchOverlayStyle: CommonVectorStyleOptions = {};
   public searchOverlayStyleSelection: CommonVectorStyleOptions = {};
   public searchOverlayStyleFocus: CommonVectorStyleOptions = {};
@@ -52,7 +67,8 @@ export class SearchState {
     private searchSourceService: SearchSourceService,
     private storageService: StorageService,
     private workspaceState: WorkspaceState,
-    private configService: ConfigService) {
+    private configService: ConfigService,
+    private mapState: MapState) {
     const searchOverlayStyle = this.configService.getConfig('searchOverlayStyle') as {
       base?: CommonVectorStyleOptions,
       selection?: CommonVectorStyleOptions,
@@ -77,6 +93,20 @@ export class SearchState {
       this.searchSourceService.setWorkspaces(wksSource, searchableWks);
     }
     );
+    this.monitorLayerDeletion();
+  }
+
+  private monitorLayerDeletion() {
+    this.mapState.map.layers$
+      .subscribe((layers) => {
+        this.searchLayerStores.forEach((store) => {
+          let layer = layers.find(l => l.id === store.layer.id);
+          if (!layer) {
+            const index = this.searchLayerStores.indexOf(store);
+            this.searchLayerStores.splice(index, 1);
+          }
+        });
+      });
   }
 
   private createCustomFilterTermStrategy(): EntityStoreFilterCustomFuncStrategy {
